+++ auto-installer_docker/run_docker.sh
@@ -1,7 +1,41 @@
-echo Killing old docker processes
+echo "*** Checking pre-requisites ***"
+# check pre-requisites: docker-compose
+which docker-compose >/dev/null
+if [ $? -ne 0 ]
+then
+    echo "[ERROR] Missing docker-compose binary - install the package and re-run this script. Aborting."
+    exit 1
+fi
+# check pre-requisites: python3
+which python3 >/dev/null
+if [ $? -ne 0 ]
+then
+    echo "[ERROR] Missing Python binary - install the package and re-run this script. Aborting."
+    exit 2
+fi
+# check pre-requisites: python3
+pip3 show netifaces >/dev/null
+if [ $? -ne 0 ]
+then
+    echo "[ERROR] Missing Python netifaces library - install and re-run this script. Aborting."
+    exit 3
+fi
+# check pre-requisites: python3
+pip3 show jinja2 >/dev/null
+if [ $? -ne 0 ]
+then
+    echo "[ERROR] Missing Python jinja2 library - install and re-run this script. Aborting."
+    exit 4
+fi
+echo "[INFO] All pre-requisites met."
+
+# (re)building docker containers
+echo
+echo "*** Killing old docker processes ***"
 docker-compose rm -fs
 
-echo Exporting host network settings
+echo
+echo "*** Exporting host network settings ***"
 
 NETWORK_DATA=( `python3 get_host_network_settings.py` )
 export EAI_HOST_IP=${NETWORK_DATA[0]}
@@ -16,5 +50,9 @@ echo EAI_HOST_SUBNET: $EAI_HOST_SUBNET
 echo EAI_HOST_NETMASK: $EAI_HOST_NETMASK
 echo
 
-echo Building docker containers
+echo "*** Checking/creating EAIDB database ***"
+python3 create_eaidb.py 
+echo
+
+echo "*** Building docker containers ***"
 docker-compose up --build -d